made search funtion case insensitive, made it so users can only delete their posts

+++ src/CocktailCard.js
@@ -3,11 +3,12 @@ import React from "react";
 function CocktailCard({id, base, description, username, image, dateCreated, onDelete }){
 
     function handleDelete(){
+        if (username === "ss"){
             fetch(`http://localhost:3000/cocktails/${id}`, {
                 method:"DELETE"
             })
             onDelete(id)
-        
+        }
     }
 
     return(
